requirements.txt
langchain
flask
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-community
langchain-huggingface
langchain-core
langchain-groq
ipykernel
-e .